Refactor: split internal/api/handlers.go into per-concern files (#66)
Splits the ~400-line handlers.go into six focused files within package api:
handler.go (core wiring), channels.go, guide.go, search.go, rss.go, categories.go.
No behaviour changes.
