+++ internal/api/categories.go
@@ -1,0 +1,12 @@
+package api
+
+import "net/http"
+
+func (h *Handler) getCategories(w http.ResponseWriter, r *http.Request) {
+	cats, err := h.db.GetCategories()
+	if err != nil {
+		http.Error(w, "database error", http.StatusInternalServerError)
+		return
+	}
+	writeJSON(w, cats)
+}
