+++ internal/api/guide.go
@@ -1,0 +1,31 @@
+package api
+
+import (
+	"net/http"
+	"time"
+)
+
+func (h *Handler) getGuide(w http.ResponseWriter, r *http.Request) {
+	dateStr := r.URL.Query().Get("date")
+	var date time.Time
+	if dateStr == "" {
+		date = time.Now()
+	} else {
+		var err error
+		date, err = time.ParseInLocation("2006-01-02", dateStr, time.Local)
+		if err != nil {
+			http.Error(w, "invalid date, expected YYYY-MM-DD", http.StatusBadRequest)
+			return
+		}
+	}
+	airings, err := h.db.GetAirings(date)
+	if err != nil {
+		http.Error(w, "database error", http.StatusInternalServerError)
+		return
+	}
+	writeJSON(w, airings)
+}
+
+func (h *Handler) getStatus(w http.ResponseWriter, r *http.Request) {
+	writeJSON(w, h.db.GetStatus())
+}
